Rename classes package to application
Closes gh-20526
