+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/java/org/springframework/boot/gradle/tasks/bundling/LayerConfiguration.java
@@ -23,10 +23,10 @@ import java.util.List;
 import org.gradle.api.Action;
 import org.gradle.api.tasks.Input;
 
-import org.springframework.boot.loader.tools.layer.classes.FilteredResourceStrategy;
-import org.springframework.boot.loader.tools.layer.classes.LocationFilter;
-import org.springframework.boot.loader.tools.layer.classes.ResourceFilter;
-import org.springframework.boot.loader.tools.layer.classes.ResourceStrategy;
+import org.springframework.boot.loader.tools.layer.application.FilteredResourceStrategy;
+import org.springframework.boot.loader.tools.layer.application.LocationFilter;
+import org.springframework.boot.loader.tools.layer.application.ResourceFilter;
+import org.springframework.boot.loader.tools.layer.application.ResourceStrategy;
 import org.springframework.boot.loader.tools.layer.library.CoordinateFilter;
 import org.springframework.boot.loader.tools.layer.library.FilteredLibraryStrategy;
 import org.springframework.boot.loader.tools.layer.library.LibraryFilter;
