+++ spring-boot-project/spring-boot-tools/spring-boot-loader-tools/src/main/java/org/springframework/boot/loader/tools/layer/application/ResourceStrategy.java
@@ -14,29 +14,26 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.loader.tools.layer.classes;
+package org.springframework.boot.loader.tools.layer.application;
 
-import java.util.List;
+import java.io.Serializable;
 
-import org.springframework.util.AntPathMatcher;
+import org.springframework.boot.loader.tools.Layer;
 
 /**
- * An implementation of {@link ResourceFilter} based on the resource location.
+ * A strategy used to match a resource to a layer.
  *
  * @author Madhura Bhave
  * @since 2.3.0
  */
-public class LocationFilter extends AbstractResourceFilter {
+public interface ResourceStrategy extends Serializable {
 
-	private static final AntPathMatcher MATCHER = new AntPathMatcher();
-
-	public LocationFilter(List<String> includes, List<String> excludes) {
-		super(includes, excludes);
-	}
-
-	@Override
-	protected boolean isMatch(String resourceName, List<String> toMatch) {
-		return toMatch.stream().anyMatch((pattern) -> MATCHER.match(pattern, resourceName));
-	}
+	/**
+	 * Return a {@link Layer} for the given resource. If no matching layer is found,
+	 * {@code null} is returned.
+	 * @param resourceName the name of the resource
+	 * @return the matching layer or {@code null}
+	 */
+	Layer getMatchingLayer(String resourceName);
 
 }
